details screen, return back to the list of articles button

+++ app/src/main/java/com/example/newsapp/domain/model/articleDetailResponse/Info.kt
@@ -20,7 +20,7 @@ data class Info(
     @SerializedName("dataType")
     val dataType: String,
     @SerializedName("sim")
-    val sim: Int,
+    val sim: Double,
     @SerializedName("url")
     val url: String,
     @SerializedName("title")
